Move Trait class to src/traits/
