+++ src/traits/Trait.js
@@ -1,0 +1,5 @@
+/**
+ * Trait interface
+ * @interface
+ */
+export class Trait {}
